% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\docType{data}
\name{legacy_raw_yield}
\alias{legacy_raw_yield}
\title{Raw crop yield data (legacy from v.0.1.1)}
\format{A data.frame with 4 variables: \code{longitude}, \code{latitude},
\code{moisture}, \code{crop}, \code{altitude}, \code{dry_bu_ac}
(dry bushels per acre), \code{watershed}, and \code{year}}
\usage{
legacy_raw_yield
}
\description{
Raw crop yield
}
\seealso{
\code{\link{legacy_yield}}
}
\keyword{datasets}
