% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\docType{data}
\name{legacy_yield_conversion}
\alias{legacy_yield_conversion}
\title{Crop yield conversions (legacy from v.0.1.1)}
\format{A data.frame with two variables: \code{crop}, \code{lbs_per_bushel},
 and \code{Mgha_per_buac}.}
\usage{
legacy_yield_conversion
}
\description{
Conversions from bushels/acre to Mg/ha for corn and soybean
}
\examples{
# Convert from bushels per acre to Megagrams per hectare
library(dplyr)
legacy_yield <- legacy_yield \%>\%
  left_join(legacy_yield_conversion, by="crop") \%>\%
  mutate(yield_Mgha = dryyield_buac * Mgha_per_buac)
}
\seealso{
\code{\link{legacy_yield}}
}
\keyword{datasets}
