% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnas_data.R
\name{pnas_data}
\alias{pnas_data}
\title{Construct water quality data used in PNAS paper}
\usage{
pnas_data()
}
\value{
A \code{data.frame} containing the columns PI, source, watershed,
year, response, and value.
}
\description{
This function will summarize surface and ground water by watershed and
year combination for analysis in the PNAS paper.
}
\examples{
d <- pnas_data()
summary(d)

}
