% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_yield}
\alias{raw_yield}
\title{Raw crop yield data}
\format{A data.frame with 4 variables: \code{longitude}, \code{latitude},
\code{moisture}, \code{crop}, \code{altitude}, \code{dry_bu_ac}
(dry bushels per acre), \code{watershed}, and \code{year}}
\usage{
raw_yield
}
\description{
Raw crop yield
}
\seealso{
\code{\link{yield}}
}
\keyword{datasets}

