% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{STRIPSyield-package}
\alias{STRIPSyield}
\alias{STRIPSyield-package}
\title{STRIPSyield: STRIPS yield data}
\description{
Crop yield data for STRIPS watersheds including observations for
    both soybean and corn years.
}
\details{
An R package containing STRIPS yield data.
To access the dataset from the PNAS paper [1], please run `pnas_data()`.
To learn more about the dataset, see the data curation vignette: `vignette('data-curation')`.
[1] Lisa A. Schulte, Jarad B. Niemi, Matthew J. Helmers, Matt Liebman, J. G. Arbuckle, David E. James, Randall K. Kolka, Matthew E. O’Neal, Mark D. Tomer, John C. Tyndall, Heidi Asbjornsen, Pauline Drobney, Jeri Neal, Gary Van Ryswyk, and Chris Witte. (2017) “Prairie strips improve biodiversity and the delivery of multiple ecosystem services from corn-soybean croplands” Proceedings of the National Academy of Sciences, 114(42), 11247-11252. http://www.pnas.org/content/114/42/11247.short
}
\author{
\strong{Maintainer}: Jarad Niemi \email{niemi@iastate.edu}

Authors:
\itemize{
  \item Matthew Liebman \email{mliebman@iastate.edu}
}

}
\keyword{internal}
