% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yield}
\alias{yield}
\title{Crop yield data}
\format{A data.frame with 4 variables: \code{year}, \code{watershed},
\code{crop}, and \code{dryyield_buac} (bushels per acre)}
\usage{
yield
}
\description{
Average crop yield for all watersheds from 2007-2016
}
\examples{
Convert from bushels per acre to Megagrams per hectare
library(dplyr)
yield <- yield \%>\%
  left_join(yield_conversion, by="crop") \%>\%
  mutate(yield_Mgha = dryyield_buac * Mgha_per_buac)
}
\seealso{
\code{\link{yield_conversion}}
}
\keyword{datasets}

