% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yield_conversion}
\alias{yield_conversion}
\title{Crop yield conversions}
\format{A data.frame with two variables: \code{crop}, \code{lbs_per_bushel},
 and \code{Mgha_per_buac}.}
\usage{
yield_conversion
}
\description{
Conversions from bushels/acre to Mg/ha for corn and soybean
}
\examples{
Convert from bushels per acre to Megagrams per hectare
library(dplyr)
yield <- yield \%>\%
  left_join(yield_conversion, by="crop") \%>\%
  mutate(yield_Mgha = dryyield_buac * Mgha_per_buac)
}
\seealso{
\code{\link{yield}}
}
\keyword{datasets}

